Remove reliance on the current buffer when opening the *org-roam* window
If the `org-roam-buffer--get-create' function is called from a hook
then `get-buffer-window' will return `nil' since the current buffer
may not have been placed in a window yet.  This eventually leads to an
error.

This change moves some code around so that we can use
`save-selected-window' to allow temporary focus on the org-roam
side window so the resizing functions can work.

I also added another window parameter: `no-other-window'.  I'm not
100% sure about this because I don't have a good feel for how the
side window is used.  I want `other-window' to avoid it but I'm not
sure if others do.